requirements.txt
opencv-python
numpy
Pillow
reportlab
streamlit
